get local repo working on Gentoo
There is one caveat, which is that when installing packages from local repo
it won't update to latest version if there is an older version already
installed. The 'update' parameter for the portage module seems to be broken
on the current version of ansible (current version being 2.0.2.0)

  failed: [localhost] (item=hello) => {"failed": true, "item": "hello",
                     "msg": "value of update must be one of: yes, got: True"}

so 'update: yes' is commented out for now.